+++ pybop/models/empirical/base_ecm.py
@@ -1,3 +1,6 @@
+import numpy as np
+import pybamm
+
 from pybop.models.base_model import BaseModel, Inputs
 from pybop.parameters.parameter_set import ParameterSet
 
@@ -77,7 +80,7 @@ class ECircuitModel(BaseModel):
         self._spatial_methods = (
             spatial_methods or self.pybamm_model.default_spatial_methods
         )
-        self.solver = solver or self.pybamm_model.default_solver
+        self._solver = solver or self.pybamm_model.default_solver
 
         # Internal attributes for the built model are initialized but not set
         self._model_with_set_params = None
@@ -100,7 +103,7 @@ class ECircuitModel(BaseModel):
         ----------
         inputs : Inputs
             The input parameters for the simulation.
-        parameter_set : pybop.parameter_set
+        parameter_set : pybop.ParameterSet
             A PyBOP parameter set object or a dictionary containing the parameter values.
         allow_infeasible_solutions : bool, optional
             If True, infeasible parameter values will be allowed in the optimisation (default: True).
@@ -111,3 +114,86 @@ class ECircuitModel(BaseModel):
             A boolean which signifies whether the parameters are compatible.
         """
         return True
+
+    def get_initial_state(
+        self,
+        initial_value,
+        parameter_values=None,
+        param=None,
+        options=None,
+        tol=1e-6,
+        inputs=None,
+    ):
+        """
+        Calculate the initial state of charge given an open-circuit voltage, voltage limits
+        and the open-circuit voltage function defined by the parameter set.
+
+        Parameters
+        ----------
+        initial_value : float
+            Target initial value.
+            If float, interpreted as SOC, must be between 0 and 1.
+            If string e.g. "4 V", interpreted as voltage, must be between V_min and V_max.
+        parameter_values : :class:`pybamm.ParameterValues`
+            The parameter values class that will be used for the simulation. Required for
+            calculating appropriate initial stoichiometries.
+        param : :class:`pybamm.LithiumIonParameters`, optional
+            The symbolic parameter set to use for the simulation.
+            If not provided, the default parameter set will be used.
+        options : dict-like, optional
+            A dictionary of options to be passed to the model, see
+            :class:`pybamm.BatteryModelOptions`.
+        tol : float, optional
+            The tolerance for the solver used to compute the initial stoichiometries.
+            A lower value results in higher precision but may increase computation time.
+            Default is 1e-6.
+
+        Returns
+        -------
+        initial_soc
+            The initial state of charge
+        """
+        parameter_values = parameter_values or self._unprocessed_parameter_set
+        param = param or self.pybamm_model.param
+
+        if isinstance(initial_value, str) and initial_value.endswith("V"):
+            V_init = float(initial_value[:-1])
+            V_min = parameter_values.evaluate(param.voltage_low_cut)
+            V_max = parameter_values.evaluate(param.voltage_high_cut)
+
+            if not V_min <= V_init <= V_max:
+                raise ValueError(
+                    f"Initial voltage {V_init}V is outside the voltage limits "
+                    f"({V_min}, {V_max})"
+                )
+
+            # Solve simple model for initial soc based on target voltage
+            soc_model = pybamm.BaseModel()
+            soc = pybamm.Variable("soc")
+            ocv = param.ocv
+            soc_model.algebraic[soc] = ocv(soc) - V_init
+
+            # initial guess for soc linearly interpolates between 0 and 1
+            # based on V linearly interpolating between V_max and V_min
+            soc_model.initial_conditions[soc] = (V_init - V_min) / (V_max - V_min)
+            soc_model.variables["soc"] = soc
+            parameter_values.process_model(soc_model)
+            initial_soc = (
+                pybamm.AlgebraicSolver(tol=tol).solve(soc_model, [0])["soc"].data[0]
+            )
+
+            # Ensure that the result lies between 0 and 1
+            initial_soc = np.minimum(np.maximum(initial_soc, 0.0), 1.0)
+
+        elif isinstance(initial_value, (int, float)):
+            if not 0 <= initial_value <= 1:
+                raise ValueError("Initial SOC should be between 0 and 1")
+            initial_soc = initial_value
+
+        else:
+            raise ValueError(
+                "Initial value must be a float between 0 and 1, "
+                "or a string ending in 'V'"
+            )
+
+        return initial_soc